# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: |-
    REST API для управления банковскими картами с JWT-аутентификацией и ролевым доступом.

    ## Аутентификация
    * Регистрация через `POST /api/auth/register` — возвращает JWT токен
    * Вход через `POST /api/auth/login` — возвращает JWT токен
    * Все защищённые endpoints требуют заголовок `Authorization: Bearer <token>`

    ## Уровни доступа
    * `ROLE_USER` — просмотр своих карт, переводы между своими картами, запрос на блокировку
    * `ROLE_ADMIN` — создание/удаление карт, управление пользователями, обработка заявок на блокировку

    ## Формат ответов
    Все endpoints возвращают `CommonResponse<T>`:
    ```json
    {
      "data": {},
      "status": "SUCCESS",
      "timestamp": "2026-04-03T22:18:39"
    }
    ```

    Ответ с ошибкой:
    ```json
    {
      "status": "ERROR",
      "problemDetail": {
        "type": "about:blank",
        "title": "Not Found",
        "status": 404,
        "detail": "Card not found"
      },
      "timestamp": "2026-04-03T22:18:39"
    }
    ```

    ## Коды ошибок
    * `400` — ошибка валидации или неверный формат запроса
    * `401` — токен отсутствует или истёк
    * `403` — недостаточно прав или аккаунт заблокирован
    * `404` — карта или пользователь не найден
    * `409` — конфликт (пользователь уже существует, заявка на блокировку уже создана)
    * `422` — недостаточно средств для перевода
    * `500` — внутренняя ошибка сервера
  contact:
...
servers:
...
security:
- Bearer Authentication: []
...
paths:
...
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Регистрация пользователя
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "409":
          description: Пользователь с таким email уже существует
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "400":
          description: Ошибка валидации
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "201":
          description: Пользователь создан — возвращает JWT токен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseAuthResponse"
      security: []
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Вход в систему
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "403":
          description: Аккаунт заблокирован или удалён
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "400":
          description: Ошибка валидации
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "401":
          description: Неверный email или пароль
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseVoid"
        "200":
          description: Успешно — возвращает JWT токен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommonResponseAuthResponse"
      security: []
...
components:
  schemas:
    CommonResponseVoid:
      type: object
      description: Общий формат ответа API
      properties:
        data:
          description: Данные ответа
        status:
          type: string
          description: Статус ответа
          enum:
          - SUCCESS
          - ERROR
          example: SUCCESS
        problemDetail:
          $ref: "#/components/schemas/ProblemDetail"
          description: Детали ошибки (только при ERROR)
        timestamp:
          type: string
          format: date-time
          description: Время ответа
          example: 2026-04-03T22:18:39
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
...
    RegisterRequest:
      type: object
      description: Регистрация пользователя в системе
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@gmail.com
          minLength: 1
        password:
          type: string
          description: Пароль
          example: password123
          maxLength: 32
          minLength: 8
      required:
      - email
      - password
    AuthResponse:
      type: object
      description: Ответ с JWT-token
      properties:
        token:
          type: string
          description: JWT токен
          example: eyJhbGciOiJIUzM4NCJ9...
    CommonResponseAuthResponse:
      type: object
      description: Общий формат ответа API
      properties:
        data:
          $ref: "#/components/schemas/AuthResponse"
          description: Данные ответа
        status:
          type: string
          description: Статус ответа
          enum:
          - SUCCESS
          - ERROR
          example: SUCCESS
        problemDetail:
          $ref: "#/components/schemas/ProblemDetail"
          description: Детали ошибки (только при ERROR)
        timestamp:
          type: string
          format: date-time
          description: Время ответа
          example: 2026-04-03T22:18:39
    LoginRequest:
      type: object
      description: Запрос на вход в систему
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@gmail.com
          minLength: 1
        password:
          type: string
          description: Пароль
          example: password123
          maxLength: 32
          minLength: 8
      required:
      - email
      - password
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: "JWT токен, полученный через POST /api/auth/login"
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT